Add hooks for before/after Model invocation
Now callers can subscribe and modify responses before they're sent to the model for processing.  The only writable property is tool_configs, which is the list of tool specs for the model.  I'm explicitly not passing toolChoice as it's hardcoded to "auto" right now and once we support actually adjusting it we can expose it as a hook.

We also don't support structured output as the parameters are not applicable